Moved Legacy content to AuthSocket.Legacy.cpp
